manifest: update the interface descriptor with feature and module_size
Add the new fields to struct greybus_descriptor_interface for the
feature bitfield and module_size enumerated type as documented in the
recent greybus specification changes.

Also add the logic to properly parse these new fields, and warn if the
features bits are set to anything we don't currently understand.

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>